Portfolio/IBKR_Trading_Engine/strategy/requirements.txt
requests==2.32.3
python-dotenv
websocket-client
numpy
tradingview-screener